Fixed split size bug

+++ src/models/action_encoder.py
@@ -41,6 +41,9 @@ class ActionEncoder(nn.Module):
         action_layers.append(nn.Sigmoid())  # Domain [0,1] for BCELoss .
         self.action = nn.Sequential(*action_layers)
 
+        # Make float
+        self.float()
+
     def forward(self, x):
         x = self.encoder(x)
 
@@ -111,13 +114,13 @@ class ActionEncoder(nn.Module):
                 kaiming_weights = torch.rand(weights.shape[0], input - weights.shape[1])
                 torch.nn.init.kaiming_uniform_(kaiming_weights, mode='fan_in', nonlinearity='leaky_relu')
 
-                weights = torch.cat([weights.double(), kaiming_weights.double()], dim=1)
+                weights = torch.cat([weights.float(), kaiming_weights.float()], dim=1)
 
             if output != weights.shape[0]:
                 kaiming_weights = torch.rand(output - weights.shape[0], input)
                 torch.nn.init.kaiming_uniform_(kaiming_weights, mode='fan_in', nonlinearity='leaky_relu')
 
-                weights = torch.cat([weights.double(), kaiming_weights.double()], dim=0)
+                weights = torch.cat([weights.float(), kaiming_weights.float()], dim=0)
 
             # Set
             layer.weight = nn.Parameter(weights)
@@ -143,7 +146,7 @@ class ActionEncoder(nn.Module):
 
             if output != biases.shape[0]:
                 rand_biases = torch.rand(output - biases.shape[0])
-                biases = torch.cat([biases, rand_biases], dim=0)
+                biases = torch.cat([biases.float(), rand_biases.float()], dim=0)
 
             # Set
             layer.bias = nn.Parameter(biases)
